Rename: Modify folder name template -> templates - modulesフォルダを作成しながら発見して修正
